+++ package.json
@@ -10,8 +10,10 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "@google-cloud/local-auth": "^2.1.0",
     "date-fns": "^4.1.0",
     "firebase": "^12.1.0",
+    "googleapis": "^105.0.0",
     "lucide-react": "^0.344.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
