--- src/home.tsx
@@ -1,1 +1,0 @@
-// tbd
